+++ orchestration/src/main/java/com/sap/ai/sdk/orchestration/SystemMessage.java
@@ -2,13 +2,16 @@ package com.sap.ai.sdk.orchestration;
 
 import static com.sap.ai.sdk.orchestration.model.SystemChatMessage.RoleEnum.SYSTEM;
 
+import com.sap.ai.sdk.orchestration.model.CacheControl;
 import com.sap.ai.sdk.orchestration.model.ChatMessage;
 import com.sap.ai.sdk.orchestration.model.ChatMessageContent;
 import com.sap.ai.sdk.orchestration.model.SystemChatMessage;
 import com.sap.ai.sdk.orchestration.model.TextContent;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.function.Function;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -36,7 +39,20 @@ public class SystemMessage implements Message {
    */
   @Tolerate
   public SystemMessage(@Nonnull final String message) {
-    content = new MessageContent(List.of(new TextItem(message)));
+    this(message, null);
+  }
+
+  /**
+   * Creates a new system message from a string, allows for cache checkpoint configuration
+   *
+   * @since 1.23.0
+   * @param message the first message.
+   * @param cacheControl prompt caching configuration to use, nullable
+   */
+  public SystemMessage(
+      @Nonnull final String message,
+      @Nullable final com.sap.ai.sdk.orchestration.CacheControl cacheControl) {
+    content = new MessageContent(List.of(new TextItem(message, cacheControl)));
   }
 
   /**
@@ -48,15 +64,54 @@ public class SystemMessage implements Message {
    */
   @Nonnull
   public SystemMessage withText(@Nonnull final String message) {
+    return withText(message, null);
+  }
+
+  /**
+   * Add text to the message
+   *
+   * @since 1.23.0
+   * @param message the text to add
+   * @param cacheControl optional cache checkpoint configuration
+   * @return the new message
+   */
+  @SuppressWarnings(
+      "PMD.PublicApiExposesModelType") // false positive: the two CacheControl classes are mixed up
+  // by the rule
+  @Nonnull
+  public SystemMessage withText(
+      @Nonnull final String message,
+      @Nullable final com.sap.ai.sdk.orchestration.CacheControl cacheControl) {
     final var contentItems = new LinkedList<>(content.items());
-    contentItems.add(new TextItem(message));
+    contentItems.add(new TextItem(message, cacheControl));
     return new SystemMessage(new MessageContent(contentItems));
   }
 
   @Nonnull
   @Override
   public ChatMessage createChatMessage() {
+    final Function<TextItem, TextContent> toTextContent =
+        (item) -> {
+          final var convertedItem =
+              TextContent.create().type(TextContent.TypeEnum.TEXT).text(item.text());
+          final var cacheControl = item.getCacheControl();
+          if (cacheControl != null) {
+            final var cacheControlConverted =
+                com.sap.ai.sdk.orchestration.model.CacheControl.create()
+                    .type(com.sap.ai.sdk.orchestration.model.CacheControl.TypeEnum.EPHEMERAL)
+                    .ttl(CacheControl.TtlEnum.fromValue(cacheControl.getTtl()));
+            convertedItem.setCacheControl(cacheControlConverted);
+          }
+          return convertedItem;
+        };
     if (content.items().size() == 1 && content.items().get(0) instanceof TextItem textItem) {
+      if (textItem.getCacheControl() != null) {
+        return SystemChatMessage.create()
+            .role(SYSTEM)
+            .content(
+                ChatMessageContent.createListOfTextContents(
+                    List.of(toTextContent.apply(textItem))));
+      }
       return SystemChatMessage.create()
           .role(SYSTEM)
           .content(ChatMessageContent.create(textItem.text()));
@@ -65,7 +120,7 @@ public class SystemMessage implements Message {
         content.items().stream()
             .filter(item -> item instanceof TextItem)
             .map(item -> (TextItem) item)
-            .map(item -> TextContent.create().type(TextContent.TypeEnum.TEXT).text(item.text()))
+            .map(toTextContent)
             .toList();
     return SystemChatMessage.create()
         .role(SYSTEM)